@@ -23,15 +23,22 @@
 
     <!-- ================= Default Properties ================= -->
     <properties>
-        <restlet.version>2.3.12</restlet.version>
-        <solr.version>4.7.2</solr.version>
+        <restlet.version>2.4.3</restlet.version>
+        <solr.version>8.7.0</solr.version>
         <sesame.version>2.7.16</sesame.version>
-        <slf4j.version>1.7.26</slf4j.version>
-        <servlet.version>3.0.1</servlet.version>
+        <slf4j.version>1.7.30</slf4j.version>
+        <log4j.version>2.14.0</log4j.version>
+        <servlet.version>3.1.0</servlet.version>
+        <xerces.version>2.9.1</xerces.version>
+        <jaxb.version>2.3.1</jaxb.version>
+        <guava.version>30.1-jre</guava.version>
+        <jetty.version>9.4.35.v20201120</jetty.version>
+        <jsonorg.version>20201115</jsonorg.version>
         <site.url>https://entrystore.org</site.url>
         <distribution.snapshot.url>https://entrystore.org</distribution.snapshot.url>
         <distribution.url>https://entrystore.org</distribution.url>
         <license.licenseName>apache_v2</license.licenseName>
+        <maven.compiler.release>11</maven.compiler.release>
     </properties>
 
     <!-- ================= Developers ================= -->
