@@ -18,14 +18,15 @@ package org.entrystore.rest.auth;
 
 import org.entrystore.Entry;
 import org.entrystore.PrincipalManager;
+import org.entrystore.config.Config;
 import org.entrystore.repository.RepositoryManager;
 import org.entrystore.repository.config.Settings;
 import org.entrystore.repository.security.Password;
+import org.entrystore.rest.filter.CORSFilter;
 import org.restlet.Request;
 import org.restlet.Response;
 import org.restlet.data.Cookie;
 import org.restlet.data.CookieSetting;
-import org.restlet.engine.util.DateUtils;
 import org.restlet.security.Verifier;
 import org.restlet.util.Series;
 import org.slf4j.Logger;
@@ -33,7 +34,6 @@ import org.slf4j.LoggerFactory;
 
 import java.net.URI;
 import java.util.Date;
-import java.util.concurrent.TimeUnit;
 
 
 /**
@@ -41,15 +41,43 @@ import java.util.concurrent.TimeUnit;
  */
 public class CookieVerifier implements Verifier {
 
-	private PrincipalManager pm;
+	private final PrincipalManager pm;
 
-	private RepositoryManager rm;
+	private final RepositoryManager rm;
 
-	private Logger log = LoggerFactory.getLogger(CookieVerifier.class);
+	private final CORSFilter corsFilter;
+
+	private static CustomCookieSettings cookieSettings;
+
+	private static final int DEFAULT_MAX_AGE = 24 * 3600;
+
+	private static final Logger log = LoggerFactory.getLogger(CookieVerifier.class);
 
 	public CookieVerifier(RepositoryManager rm) {
+		this(rm, null);
+	}
+
+	public CookieVerifier(RepositoryManager rm, CORSFilter corsFilter) {
 		this.rm = rm;
 		this.pm = rm.getPrincipalManager();
+		this.corsFilter = corsFilter;
+		if (cookieSettings == null) {
+			Config config = rm.getConfiguration();
+			CustomCookieSettings.SameSite sameSite = CustomCookieSettings.SameSite.Strict;
+			try {
+				String sameSiteStr = config.getString(Settings.AUTH_COOKIE_SAMESITE, CustomCookieSettings.SameSite.Strict.name()).toLowerCase();
+				if (sameSiteStr.length() > 1) {
+					sameSiteStr = sameSiteStr.substring(0, 1).toUpperCase() + sameSiteStr.substring(1);
+				}
+				sameSite = CustomCookieSettings.SameSite.valueOf(sameSiteStr);
+			} catch (IllegalArgumentException iae) {
+				log.warn("Invalid value for setting " + Settings.AUTH_COOKIE_SAMESITE + ": " + iae.getMessage());
+			}
+			cookieSettings = new CustomCookieSettings(
+					config.getBoolean(Settings.AUTH_COOKIE_SECURE, true),
+					config.getBoolean(Settings.AUTH_COOKIE_HTTPONLY, true),
+					sameSite);
+		}
 	}
 
 	public int verify(Request request, Response response) {
@@ -77,11 +105,17 @@ public class CookieVerifier implements Verifier {
 					} else {
 						log.error("Auth token maps to non-existing user, removing token");
 						tc.removeToken(authToken);
-						cleanCookies("auth_token", request, response);
+						cleanCookies(rm,"auth_token", request, response);
 					}
 				} else {
-					log.debug("Auth token not found in token cache: " + authToken);
-					cleanCookies("auth_token", request, response);
+					log.debug("Auth token not found in token cache");
+					cleanCookies(rm,"auth_token", request, response);
+					// CORS needs to be handled here, because we return a RESULT_INVALID which
+					// interrupts the filter chain before the CORS filter can do its work
+					if (corsFilter != null) {
+						corsFilter.addCorsHeader(request, response);
+					}
+					// return RESULT_INVALID;
 				}
 			}
 
@@ -95,115 +129,60 @@ public class CookieVerifier implements Verifier {
 		}
 	}
 
-	public static void cleanCookies(String cookieName, Request request, Response response) {
+	public static void cleanCookies(RepositoryManager rm, String cookieName, Request request, Response response) {
 		response.getCookieSettings().removeAll(cookieName);
 		Series<Cookie> cookies = request.getCookies();		
 		for (Cookie c : cookies) {
 			if (c.getName().equals(cookieName)) {
-				CookieSetting cs = new CookieSetting(c.getVersion(), c.getName(), c.getValue());
-				cs.setPath(c.getPath());
-				// we don't use Restlet's getCookieSettings().add(cs) with
-				// Max-Age of 0 due to some shortcoming of the implementation,
-				// see comment on createExpireCookie() below
-				response.getHeaders().add("Set-Cookie", createExpiredCookie(cs));
+				// The following is a hack, explained in createAuthToken() below
+				String value = c.getValue() + "; Max-Age=0";
+				CookieSetting cs = new CookieSetting(c.getVersion(), c.getName(), value, getCookiePath(rm), null);
+				cs.setMaxAge(0);
+				response.getCookieSettings().add(cs);
 			}
 		}
 	}
 
-	/**
-	 * This method is a workaround for Restlet's inability of setting a Cookie
-	 * expiration date in the past. This is needed for some browsers (e.g. Edge)
-	 * which take the local machine's time instead of the server time to decide
-	 * whether the cookie is expired or not.
-	 *
-	 * This method should only be used to unset/expire a Cookie. Also, for the
-	 * sake of simplicity it lacks support for Cookie version 1.
-	 *
-	 * @param cookieSetting The Cookie setting to be used to write the header.
-	 * @return The value of the header (without "Set-Cookie:") to be set on a Response body.
-	 */
-	private static String createExpiredCookie(CookieSetting cookieSetting) {
-		if (cookieSetting.getVersion() > 0) {
-			throw new IllegalArgumentException("This method does not support Cookie version 1");
-		}
-
-		StringBuffer result = new StringBuffer();
-
-		String name = cookieSetting.getName();
-		String value = cookieSetting.getValue();
-
-		if ((name == null) || (name.length() == 0)) {
-			throw new IllegalArgumentException(
-					"Can't write cookie. Invalid name detected");
-		}
-
-		result.append(name).append('=');
-
-		// Append the value
-		if ((value != null) && (value.length() > 0)) {
-			result.append(value);
-		}
-
-		// Append the path
-		String path = cookieSetting.getPath();
-
-		if ((path != null) && (path.length() > 0)) {
-			result.append("; Path=").append(path);
-		}
-
-		// Append the expiration date
-		long currentTime = System.currentTimeMillis();
-		long expiresTime = currentTime - TimeUnit.DAYS.toMillis(1);
-		Date expires = new Date(expiresTime);
-
-		result.append("; Expires=");
-		result.append(DateUtils.format(expires,	DateUtils.FORMAT_RFC_1123.get(0)));
-
-		// Append the domain
-		String domain = cookieSetting.getDomain();
-
-		if ((domain != null) && (domain.length() > 0)) {
-			result.append("; Domain=");
-			result.append(domain.toLowerCase());
-		}
-
-		// Append the secure flag
-		if (cookieSetting.isSecure()) {
-			result.append("; Secure");
-		}
-
-		// Append the secure flag
-		if (cookieSetting.isAccessRestricted()) {
-			result.append("; HttpOnly");
-		}
-
-		return result.toString();
-	}
-
 	public void createAuthToken(String userName, String maxAgeStr, Response response) {
+		Config config = rm.getConfiguration();
 		// 24h default, lifetime in seconds
-		int maxAge = rm.getConfiguration().getInt(Settings.AUTH_TOKEN_MAX_AGE, 24 * 3600);
+		long maxAge = config.getLong(Settings.AUTH_TOKEN_MAX_AGE, config.getLong(Settings.AUTH_COOKIE_MAXAGE, DEFAULT_MAX_AGE));
 		if (maxAgeStr != null) {
 			try {
-				maxAge = Integer.parseInt(maxAgeStr);
-			} catch (NumberFormatException nfe) {}
+				maxAge = Long.parseLong(maxAgeStr);
+			} catch (NumberFormatException ignored) {}
 		}
 
 		String token = Password.getRandomBase64(128);
-		Date loginExpiration = new Date(new Date().getTime() + (maxAge * 1000));
+		Date loginExpiration = new Date(new Date().getTime() + (DEFAULT_MAX_AGE * 1000));
+		if (maxAge >= 0) { // negative value means session cookie, see below
+			loginExpiration = new Date(new Date().getTime() + (maxAge * 1000));
+		}
 		LoginTokenCache.getInstance().putToken(token, new UserInfo(userName, loginExpiration));
-
 		log.debug("User " + userName + " receives authentication token that will expire on " + loginExpiration);
 
+		// We hack the mechanism and set additional properties as part of the Cookie value since
+		// there is no direct way to set properties such as Max-Age, SameSite, etc.
+		// This works since Restlet does not parse or process the value; this hack might break in the future.
+		// We only set Max-Age for positive values; omission of Max-Age and Expires makes it a session cookie.
+		if (maxAge >= 0) {
+			token += "; Max-Age=" + maxAge;
+		}
+		token += "; " + cookieSettings.toString();
 		CookieSetting tokenCookieSetting = new CookieSetting(0, "auth_token", token);
-		tokenCookieSetting.setMaxAge(maxAge);
+		// CookieSetting.setMaxAge() actually materializes as an "Expires" setting for some strange reason,
+		// that's why we set "Max-Age" (which takes precedent over "Expires") above instead.
+		// tokenCookieSetting.setMaxAge(maxAge);
+		tokenCookieSetting.setPath(getCookiePath(rm));
+		response.getCookieSettings().add(tokenCookieSetting);
+	}
+
+	private static String getCookiePath(RepositoryManager rm) {
 		String cookiePath = rm.getConfiguration().getString(Settings.AUTH_COOKIE_PATH, "auto");
 		if ("auto".equalsIgnoreCase(cookiePath)) {
-			tokenCookieSetting.setPath(rm.getRepositoryURL().getPath());
-		} else {
-			tokenCookieSetting.setPath(cookiePath);
+			return rm.getRepositoryURL().getPath();
 		}
-		response.getCookieSettings().add(tokenCookieSetting);
+		return cookiePath;
 	}
 
 }
