@@ -53,9 +53,11 @@ import java.net.URI;
 import java.net.URL;
 import java.net.URLDecoder;
 import java.net.URLEncoder;
+import java.nio.charset.StandardCharsets;
 import java.security.Security;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -71,6 +73,20 @@ public class SamlLoginResource extends BaseResource {
 
 	private static SamlClient samlClient;
 
+	private static String relyingPartyId;
+
+	private static String assertionConsumerService;
+
+	private static String idpMetadataUrl;
+
+	private static String redirSuccess;
+
+	private static String redirFailure;
+
+	private static Date metadataLoaded = new Date();
+
+	private static long metadataMaxAge;
+
 	static {
 		Security.addProvider(new BouncyCastleProvider());
 	}
@@ -82,42 +98,61 @@ public class SamlLoginResource extends BaseResource {
 		if (samlClient == null) {
 			Config config = getRM().getConfiguration();
 
-			String relyingPartyId = config.getString(Settings.AUTH_SAML_RELYING_PARTY_ID);
+			relyingPartyId = config.getString(Settings.AUTH_SAML_RELYING_PARTY_ID);
 			if (relyingPartyId == null) {
 				log.warn("No SAML Relying Party Identifier configured");
 			} else {
 				log.info("SAML Relying Party Identifier: " + relyingPartyId);
 			}
 
-			String assertionConsumerService = config.getString(Settings.AUTH_SAML_ASSERTION_CONSUMER_SERVICE_URL);
+			assertionConsumerService = config.getString(Settings.AUTH_SAML_ASSERTION_CONSUMER_SERVICE_URL);
 			if (assertionConsumerService == null) {
 				log.warn("No SAML Assertion Consumer Service URL configured");
 			} else {
 				log.info("SAML Assertion Consumer Service URL: " + assertionConsumerService);
 			}
 
-			String idpMetadata = config.getString(Settings.AUTH_SAML_IDP_METADATA_URL);
-			if (idpMetadata == null) {
+			idpMetadataUrl = config.getString(Settings.AUTH_SAML_IDP_METADATA_URL);
+			if (idpMetadataUrl == null) {
 				log.warn("No SAML IDP Metadata URL configured");
 			} else {
-				log.info("SAML IDP Metadata URL: " + idpMetadata);
+				log.info("SAML IDP Metadata URL: " + idpMetadataUrl);
 			}
 
-			if (relyingPartyId != null && assertionConsumerService != null && idpMetadata != null) {
-				try {
-					Reader idpMetadataReader = new BufferedReader(new InputStreamReader(new URL(idpMetadata).openStream()));
-					samlClient = SamlClient.fromMetadata(relyingPartyId, assertionConsumerService, idpMetadataReader);
-				} catch (IOException | SamlException e) {
-					log.error(e.getMessage());
-				}
+			metadataMaxAge = config.getLong(Settings.AUTH_SAML_IDP_METADATA_MAXAGE, 604800) * 1000;
+
+			redirSuccess = config.getString(Settings.AUTH_SAML_REDIRECT_SUCCESS_URL);
+			redirFailure = config.getString(Settings.AUTH_SAML_REDIRECT_FAILURE_URL);
+
+			if (relyingPartyId != null && assertionConsumerService != null && idpMetadataUrl != null) {
+				loadMetadataAndInitSamlClient();
 			} else {
 				log.error("SAML authentication could not be initialized properly");
 			}
 		}
 	}
 
+	private void loadMetadataAndInitSamlClient() {
+		try {
+			Reader idpMetadataReader = new BufferedReader(new InputStreamReader(new URL(idpMetadataUrl).openStream(), StandardCharsets.UTF_8));
+			samlClient = SamlClient.fromMetadata(relyingPartyId, assertionConsumerService, idpMetadataReader);
+			metadataLoaded = new Date();
+			log.info("Reloaded SAML metadata from " + idpMetadataUrl);
+		} catch (IOException | SamlException e) {
+			log.error(e.getMessage());
+		}
+	}
+
+	private void checkAndInitSamlClient() {
+		if ((new Date().getTime() - metadataLoaded.getTime()) > metadataMaxAge) {
+			log.info("Reloading SAML metadata");
+			loadMetadataAndInitSamlClient();
+		}
+	}
+
 	@Get
 	public Representation represent() {
+		checkAndInitSamlClient();
 		try {
 			redirectToIdentityProvider(getResponse());
 		} catch (SamlException e) {
@@ -130,11 +165,9 @@ public class SamlLoginResource extends BaseResource {
 
 	@Post
 	public void store(Representation r) {
-		String redirSuccess = parameters.get("redirectOnSuccess");
-		String redirFailure = parameters.get("redirectOnFailure");
+		checkAndInitSamlClient();
 
 		boolean html = MediaType.TEXT_HTML.equals(getRequest().getClientInfo().getPreferredMediaType(Arrays.asList(MediaType.TEXT_HTML, MediaType.APPLICATION_ALL)));
-
 		boolean authSuccess = false;
 
 		String encodedResponse = new Form(r).getFirstValue("SAMLResponse");
@@ -221,28 +254,6 @@ public class SamlLoginResource extends BaseResource {
 				log.error(e.getMessage());
 			}
 		}
-	}
-
-	private String constructServiceUrl(String redirSuccess, String redirFailure) {
-		String result = getBaseUrl() + "auth/saml";
-		if (redirSuccess != null) {
-			result += "?redirectOnSuccess=";
-			result += redirSuccess;
-		}
-		if (redirFailure != null) {
-			result += (redirSuccess == null ? "?" : "&");
-			result += "redirectOnFailure=";
-			result += redirFailure;
-		}
-		return result;
-	}
-
-	private String getBaseUrl() {
-		String baseUrl = getRM().getConfiguration().getString(Settings.BASE_URL);
-		if (!baseUrl.endsWith("/")) {
-			baseUrl += "/";
-		}
-		return baseUrl;
 	}
 
 	private void redirectToIdentityProvider(Response response) throws SamlException {
