@@ -27,6 +27,7 @@ import org.entrystore.PrincipalManager;
 import org.entrystore.User;
 import org.entrystore.config.Config;
 import org.entrystore.repository.config.Settings;
+import org.entrystore.repository.security.Password;
 import org.entrystore.rest.auth.Signup;
 import org.entrystore.rest.auth.SignupInfo;
 import org.entrystore.rest.auth.SignupTokenCache;
@@ -62,7 +63,6 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Set;
 
-
 /**
  * Resource to handle manual sign-ups.
  * 
@@ -70,13 +70,13 @@ import java.util.Set;
  */
 public class SignupResource extends BaseResource {
 	
-	private static Logger log = LoggerFactory.getLogger(SignupResource.class);
+	private static final Logger log = LoggerFactory.getLogger(SignupResource.class);
 
 	protected SimpleHTML html = new SimpleHTML("Sign-up");
 
 	private static Set<String> domainWhitelist = null;
 
-	private static Object mutex = new Object();
+	private static final Object mutex = new Object();
 
 	@Override
 	public void doInit() {
@@ -156,7 +156,7 @@ public class SignupResource extends BaseResource {
 			pm.setPrincipalName(entry.getResourceURI(), ci.email);
 			Signup.setFoafMetadata(entry, new org.restlet.security.User("", "", ci.firstName, ci.lastName, ci.email));
 			User u = (User) entry.getResource();
-			u.setSecret(ci.password);
+			u.setSaltedHashedSecret(ci.saltedHashedPassword);
 			if (ci.customProperties != null) {
 				u.setCustomProperties(ci.customProperties);
 			}
@@ -201,6 +201,7 @@ public class SignupResource extends BaseResource {
 		String rcResponse = null;
 		String rcResponseV2 = null;
 		String customPropPrefix = "custom_";
+		String password = null;
 
 		if (MediaType.APPLICATION_JSON.equals(r.getMediaType())) {
 			try {
@@ -215,7 +216,7 @@ public class SignupResource extends BaseResource {
 					ci.email = siJson.getString("email");
 				}
 				if (siJson.has("password")) {
-					ci.password = siJson.getString("password");
+					password = siJson.getString("password");
 				}
 				if (siJson.has("recaptcha_challenge_field")) {
 					rcChallenge = siJson.getString("recaptcha_challenge_field");
@@ -250,7 +251,7 @@ public class SignupResource extends BaseResource {
 			ci.firstName = form.getFirstValue("firstname", true);
 			ci.lastName = form.getFirstValue("lastname", true);
 			ci.email = form.getFirstValue("email", true);
-			ci.password = form.getFirstValue("password", true);
+			password = form.getFirstValue("password", true);
 			rcChallenge = form.getFirstValue("recaptcha_challenge_field", true);
 			rcResponse = form.getFirstValue("recaptcha_response_field", true);
 			rcResponseV2 = form.getFirstValue("g-recaptcha-response", true);
@@ -265,7 +266,7 @@ public class SignupResource extends BaseResource {
 			}
 		}
 
-		if (ci.firstName == null || ci.lastName == null || ci.email == null || ci.password == null) {
+		if (ci.firstName == null || ci.lastName == null || ci.email == null || password == null) {
 			getResponse().setStatus(Status.CLIENT_ERROR_BAD_REQUEST);
 			getResponse().setEntity(html.representation("One or more parameters are missing."));
 			return;
@@ -277,7 +278,7 @@ public class SignupResource extends BaseResource {
 			return;
 		}
 
-		if (ci.password.trim().length() < 8) {
+		if (password.trim().length() < 8) {
 			getResponse().setStatus(Status.CLIENT_ERROR_BAD_REQUEST);
 			getResponse().setEntity(html.representation("The password has to consist of at least 8 characters."));
 			return;
@@ -336,10 +337,11 @@ public class SignupResource extends BaseResource {
 
 		String token = RandomStringUtils.random(16, 0, 0, true, true, null, new SecureRandom());
 		String confirmationLink = getRM().getRepositoryURL().toExternalForm() + "auth/signup?confirm=" + token;
-		log.info("Generated sign-up token " + token + " for " + ci.email);
+		log.info("Generated sign-up token for " + ci.email);
 
 		boolean sendSuccessful = Email.sendSignupConfirmation(getRM().getConfiguration(), ci.firstName + " " + ci.lastName, ci.email, confirmationLink);
 		if (sendSuccessful) {
+			ci.saltedHashedPassword = Password.getSaltedHash(password);
 			SignupTokenCache.getInstance().putToken(token, ci);
 			log.info("Sent confirmation request to " + ci.email);
 		} else {
@@ -367,12 +369,6 @@ public class SignupResource extends BaseResource {
 			if (siteKey == null) {
 				log.warn("reCaptcha keys must be configured; rendering form without reCaptcha");
 			} else {
-				/* reCaptcha 1.0 (deprecated)
-				String publicKey = config.getString(Settings.AUTH_RECAPTCHA_PUBLIC_KEY);
-				ReCaptcha c = ReCaptchaFactory.newReCaptcha(publicKey, privateKey, false);
-				reCaptchaHtml = c.createRecaptchaHtml(null, null);
-				*/
-
 				// reCaptcha 2.0
 				sb.append("<script src=\"https://www.google.com/recaptcha/api.js\" async defer></script>\n");
 				sb.append("<p>\n<div class=\"g-recaptcha\" data-sitekey=\"").append(siteKey).append("\"></div>\n</p>\n");
@@ -380,27 +376,8 @@ public class SignupResource extends BaseResource {
 		}
 		sb.append("<br/>\n<input type=\"submit\" value=\"Sign-up\" />\n");
 		sb.append("</form>\n");
-
-		boolean openid = "on".equalsIgnoreCase(config.getString(Settings.AUTH_OPENID, "off"));
-		if (openid) {
-			boolean google = "on".equalsIgnoreCase(config.getString(Settings.AUTH_OPENID_GOOGLE, "off"));
-			boolean yahoo = "on".equalsIgnoreCase(config.getString(Settings.AUTH_OPENID_YAHOO, "off"));
-			if (google || yahoo) {
-				sb.append("<br/>\n");
-				sb.append("Sign-up with: ");
-				if (google) {
-					sb.append("<a href=\"openid/google/signup\">Google</a>\n");
-				}
-				if (yahoo) {
-					if (google) {
-						sb.append(" | ");
-					}
-					sb.append("<a href=\"openid/yahoo/signup\">Yahoo!</a>\n");
-				}
-			}
-		}
-
 		sb.append(html.footer());
+
 		return sb.toString();
 	}
 
