@@ -32,7 +32,10 @@ import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.net.URL;
-import java.util.*;
+import java.nio.charset.StandardCharsets;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 
 /**
  * @author Hannes Ebner
@@ -139,7 +142,7 @@ public class HttpUtil {
 	public static String readFirstLine(URL url) {
 		BufferedReader in = null;
 		try {
-			in = new BufferedReader(new InputStreamReader(url.openStream()));
+			in = new BufferedReader(new InputStreamReader(url.openStream(), StandardCharsets.UTF_8));
 			return in.readLine();
 		} catch (IOException ioe) {
 			log.error(ioe.getMessage());
